requirements.txt
coverage==4.5.4
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
